refactor: move Main to testclasses
